medical-monitor-jorge/requirements.txt
streamlit>=1.38.0
duckdb>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
Pillow>=10.0.0